third_party/blink/web_tests/wpt_internal/prerender/register-service-worker.https.html

<!DOCTYPE html>
<title>Registration of a new service worker in a prerendered page</title>
<meta name="timeout" content="long">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/common/utils.js"></script>
<script src="/service-workers/service-worker/resources/test-helpers.sub.js"></script>
<script src="/speculation-rules/prerender/resources/utils.js"></script>
<body>
<script>
const uid = token();

const PAGE_URL = `resources/register-service-worker.html?uid=${uid}`;

// To make sure the service worker registered by the prerendered page starts up,
// this test sends messages as the following sequence:
// prerendered page => service worker => prerendered page => main page.
promise_test(async t => {
  const bc = new PrerenderChannel('prerender-channel', uid);
  t.add_cleanup(_ => bc.close());

  const gotMessage = new Promise(resolve => {
    bc.addEventListener('message', e => {
      resolve(e.data);
    }, {
      once: true
    });
  });

  startPrerendering(PAGE_URL);

  const result = await gotMessage;
  assert_equals(result, 'pingpong');
}, 'New service worker should be registered in a prerendered page');

</script>
</body>
